final commit (?!)

+++ app/Http/Controllers/CommentController.php
@@ -2,50 +2,28 @@
 
 namespace App\Http\Controllers;
 
-use App\Notifications\CommentNotification;
-use App\Notifications\ReplyNotification;
-use Illuminate\Http\Request;
-use App\Models\NewsPost;
 use App\Models\Comment;
+use App\Models\CommentReport;
+use App\Models\NewsPost;
+use App\Notifications\CommentNotification;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Support\Facades\DB;
 
 class CommentController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function index()
-    {
-        //
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create($body, $id_post)
-    {
-        return Comment::create([
-            'body' => $body,
-            'date_time' => now(),
-            'aura' => 0,
-            'id_owner' => Auth::user()->id,
-            'id_post' => $id_post
-        ]);
-    }
-
     public function comment(Request $request, NewsPost $newspost)
     {
         if (!Auth::check()) return response()->json('Forbidden access', 403);
 
-        $comment = $this->create($request->input('comment'), $newspost->id);
+        $comment = Comment::create([
+            'body' => $request->input('comment'),
+            'date_time' => now(),
+            'aura' => 0,
+            'id_owner' => Auth::user()->id,
+            'id_post' => $newspost->id
+        ]);
         $comment->save();
-        $comment = $comment->fresh(); //refresh model
+        $comment->refresh();
         $html = view('partials.comment', ['comment' => $comment, 'offset' => 0])->render();
         $newspost->owner->notify(new CommentNotification($comment));
         return response()->json($html);
@@ -54,15 +32,15 @@ class CommentController extends Controller
 
     public function reply(Request $request, NewsPost $newspost, Comment $comment)
     {
+        if (!Auth::check()) return response()->json('Forbidden access', 403);
+
         $reply = $comment->reply_transaction($request->input('comment'), $comment->id, $newspost->id);
         $reply->save();
 
-        $newspost = $newspost->refresh();
+        $newspost->refresh();
+        $reply = $reply->refresh();
 
-        $html = [];
-        foreach ($newspost->parentComments() as $parent) {
-            array_push($html, view('partials.comment', ['comment' => $parent, 'offset' => 0])->render());
-        }
+        $html = view('partials.comment', ['comment' => $reply, 'offset' => $request->input('offset')])->render();
 
         return response()->json(array('parent_comment_id' => $comment->id, 'html' => $html));
     }
@@ -77,8 +55,7 @@ class CommentController extends Controller
 
         } else if (($vote->upvote == 1 && $request->input('vote') === 'true') || ($vote->upvote == 0 && $request->input('vote') === 'false')) {
             $comment->delete_vote();
-        } 
-        else {
+        } else {
 
             $comment->delete_vote();
             $comment->add_vote($request->input('vote'));
@@ -86,39 +63,6 @@ class CommentController extends Controller
 
         $comment->refresh();
         return response()->json(array('votes' => $comment->aura));
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
-    }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param int $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @return \Illuminate\Http\Response
-     */
-    public function edit(Request $request)
-    {
-
     }
 
     /**
@@ -130,6 +74,7 @@ class CommentController extends Controller
      */
     public function update(Request $request, Comment $comment)
     {
+        if (!Auth::check()) return response()->json('Forbidden access', 403);
         $this->authorize('owner', $comment);
         $comment->body = $request->input('body');
         $comment->save();
@@ -152,23 +97,16 @@ class CommentController extends Controller
      */
     public function destroy(Comment $comment)
     {
+        if (!Auth::check()) return response()->json('Forbidden access', 403);
         $this->authorize('delete', $comment);
 
-        $post = $comment->post;
         $comment->delete();
-        $html = [];
-
-        foreach ($post->parentComments() as $parent) {
-            array_push($html, view('partials.comment', ['comment' => $parent, 'offset' => 0])->render());
-        }
-
-        return response()->json(array('html' => $html));
     }
 
     public function report(Request $request, Comment $comment)
     {
         if (!Auth::check()) return response()->json(array('auth' => 'Forbidden Access'), 403);
-        
+
         return CommentReport::create([
             'id_reporter' => Auth::user()->id,
             'id_comment' => $comment->id,
@@ -178,6 +116,7 @@ class CommentController extends Controller
 
     public function dismiss(Comment $comment)
     {
+        if (!Auth::check() || !Auth::user()->admin) return response()->json(array('auth' => 'Forbidden Access'), 403);
         $comment->dismiss_report();
     }
 }
